Dynamically load hive and avro using reflection to avoid potential class not found exception
