+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/ExternalSource.scala
@@ -21,15 +21,46 @@ import scala.util.{Failure, Success, Try}
 import com.nvidia.spark.rapids._
 
 import org.apache.spark.broadcast.Broadcast
-import org.apache.spark.sql.avro.{AvroFileFormat, AvroOptions}
 import org.apache.spark.sql.connector.read.{PartitionReaderFactory, Scan}
 import org.apache.spark.sql.execution.FileSourceScanExec
 import org.apache.spark.sql.execution.datasources.FileFormat
 import org.apache.spark.sql.sources.Filter
-import org.apache.spark.sql.v2.avro.AvroScan
 import org.apache.spark.util.{SerializableConfiguration, Utils}
 
+/**
+ * The subclass of AvroProvider imports spark-avro classes. This file should not imports
+ * spark-avro because `class not found` exception may throw if spark-avro does not exist at
+ * runtime. Details see: https://github.com/NVIDIA/spark-rapids/issues/5648
+ */
+trait AvroProvider {
+  /** If the file format is supported as an external source */
+  def isSupportedFormat(format: FileFormat): Boolean
+
+  def isPerFileReadEnabledForFormat(format: FileFormat, conf: RapidsConf): Boolean
+
+  def tagSupportForGpuFileSourceScan(meta: SparkPlanMeta[FileSourceScanExec]): Unit
+
+  /**
+   * Get a read file format for the input format.
+   * Better to check if the format is supported first by calling 'isSupportedFormat'
+   */
+  def getReadFileFormat(format: FileFormat): FileFormat
+
+  /**
+   * Create a multi-file reader factory for the input format.
+   * Better to check if the format is supported first by calling 'isSupportedFormat'
+   */
+  def createMultiFileReaderFactory(
+      format: FileFormat,
+      broadcastedConf: Broadcast[SerializableConfiguration],
+      pushedFilters: Array[Filter],
+      fileScan: GpuFileSourceScanExec): PartitionReaderFactory
+
+  def getScans: Map[Class[_ <: Scan], ScanRule[_ <: Scan]]
+}
+
 object ExternalSource {
+  val providerClassName = "org.apache.spark.sql.rapids.AvroSourceProvider"
 
   lazy val hasSparkAvroJar = {
     val loader = Utils.getContextOrSparkClassLoader
@@ -45,29 +76,20 @@ object ExternalSource {
   /** If the file format is supported as an external source */
   def isSupportedFormat(format: FileFormat): Boolean = {
     if (hasSparkAvroJar) {
-      format match {
-        case _: AvroFileFormat => true
-        case _ => false
-      }
+      ShimLoader.newInstanceOf[AvroProvider](providerClassName).isSupportedFormat(format)
     } else false
   }
 
   def isPerFileReadEnabledForFormat(format: FileFormat, conf: RapidsConf): Boolean = {
     if (hasSparkAvroJar) {
-      format match {
-        case _: AvroFileFormat => conf.isAvroPerFileReadEnabled
-        case _ => false
-      }
+      ShimLoader.newInstanceOf[AvroProvider](providerClassName)
+          .isPerFileReadEnabledForFormat(format, conf)
     } else false
   }
 
   def tagSupportForGpuFileSourceScan(meta: SparkPlanMeta[FileSourceScanExec]): Unit = {
     if (hasSparkAvroJar) {
-      meta.wrapped.relation.fileFormat match {
-        case _: AvroFileFormat => GpuReadAvroFileFormat.tagSupport(meta)
-        case f =>
-          meta.willNotWorkOnGpu(s"unsupported file format: ${f.getClass.getCanonicalName}")
-      }
+      ShimLoader.newInstanceOf[AvroProvider](providerClassName).tagSupportForGpuFileSourceScan(meta)
     }
   }
 
@@ -77,11 +99,7 @@ object ExternalSource {
    */
   def getReadFileFormat(format: FileFormat): FileFormat = {
     if (hasSparkAvroJar) {
-      format match {
-        case _: AvroFileFormat => new GpuReadAvroFileFormat
-        case f =>
-          throw new IllegalArgumentException(s"${f.getClass.getCanonicalName} is not supported")
-      }
+      ShimLoader.newInstanceOf[AvroProvider](providerClassName).getReadFileFormat(format)
     } else {
       throw new IllegalArgumentException(s"${format.getClass.getCanonicalName} is not supported")
     }
@@ -97,50 +115,16 @@ object ExternalSource {
       pushedFilters: Array[Filter],
       fileScan: GpuFileSourceScanExec): PartitionReaderFactory = {
     if (hasSparkAvroJar) {
-      format match {
-        case _: AvroFileFormat =>
-          GpuAvroMultiFilePartitionReaderFactory(
-            fileScan.relation.sparkSession.sessionState.conf,
-            fileScan.rapidsConf,
-            broadcastedConf,
-            fileScan.relation.dataSchema,
-            fileScan.requiredSchema,
-            fileScan.relation.partitionSchema,
-            new AvroOptions(fileScan.relation.options, broadcastedConf.value.value),
-            fileScan.allMetrics,
-            pushedFilters,
-            fileScan.queryUsesInputFile)
-        case _ =>
-          // never reach here
-          throw new RuntimeException(s"File format $format is not supported yet")
-      }
+      ShimLoader.newInstanceOf[AvroProvider](providerClassName).
+          createMultiFileReaderFactory(format, broadcastedConf, pushedFilters, fileScan)
     } else {
       throw new RuntimeException(s"File format $format is not supported yet")
     }
-
   }
 
   def getScans: Map[Class[_ <: Scan], ScanRule[_ <: Scan]] = {
     if (hasSparkAvroJar) {
-      Seq(
-        GpuOverrides.scan[AvroScan](
-          "Avro parsing",
-          (a, conf, p, r) => new ScanMeta[AvroScan](a, conf, p, r) {
-            override def tagSelfForGpu(): Unit = GpuAvroScan.tagSupport(this)
-
-            override def convertToGpu(): Scan =
-              GpuAvroScan(a.sparkSession,
-                a.fileIndex,
-                a.dataSchema,
-                a.readDataSchema,
-                a.readPartitionSchema,
-                a.options,
-                a.pushedFilters,
-                conf,
-                a.partitionFilters,
-                a.dataFilters)
-          })
-      ).map(r => (r.getClassFor.asSubclass(classOf[Scan]), r)).toMap
+      ShimLoader.newInstanceOf[AvroProvider](providerClassName).getScans
     } else Map.empty
   }
